+++ client/one/loader_responses.go
@@ -9,26 +9,26 @@ type LoaderStatus struct {
 	Success     bool  `json:"success,omitempty"`
 	Loaded      bool  `json:"loaded,omitempty"`
 	Now         int64 `json:"now,omitempty"`
-	BlocksCount int   `json:"blocksCount,omitempty"`
+	BlocksCount byte  `json:"blocksCount,omitempty"`
 }
 
 type LoaderSync struct {
 	Success bool   `json:"success,omitempty"`
 	Syncing bool   `json:"syncing,omitempty"`
-	Blocks  int    `json:"blocks,omitempty"`
+	Blocks  byte   `json:"blocks,omitempty"`
 	Height  int64  `json:"height,omitempty"`
 	Id      string `json:"id,omitempty"`
 }
 
 type LoaderAutoConfigure struct {
-	Success       bool           `json:"success,omitempty"`
-	Nethash       string         `json:"nethash,omitempty"`
-	Token         string         `json:"token,omitempty"`
-	Symbol        string         `json:"symbol,omitempty"`
-	Explorer      string         `json:"explorer,omitempty"`
-	Version       int            `json:"version,omitempty"`
-	Ports         map[string]int `json:"ports,omitempty"`
-	FeeStatistics []FeeStatistic `json:"feeStatistics,omitempty"`
+	Success       bool            `json:"success,omitempty"`
+	Nethash       string          `json:"nethash,omitempty"`
+	Token         string          `json:"token,omitempty"`
+	Symbol        string          `json:"symbol,omitempty"`
+	Explorer      string          `json:"explorer,omitempty"`
+	Version       byte            `json:"version,omitempty"`
+	Ports         map[string]byte `json:"ports,omitempty"`
+	FeeStatistics []FeeStatistic  `json:"feeStatistics,omitempty"`
 }
 
 type FeeStatistic struct {
